utils/install/requirements.txt
GitPython==3.1.40
inquirer==3.1.4
termcolor==2.4.0